Cleanups

+++ src/main/java/cc/hyperium/mods/VictoryRoyale.java
@@ -107,24 +107,21 @@ public class VictoryRoyale extends AbstractMod {
                     worldRenderer.begin(GL11.GL_TRIANGLES, DefaultVertexFormats.POSITION);
 
                     if (side == 1) {
-                        worldRenderer.pos((double) trueX, (double) trueY, 0.0D).endVertex(); // x
+                        worldRenderer.pos(trueX, trueY, 0.0D).endVertex(); // x
                         worldRenderer.pos((double) px + 0, (double) py + point.width, 0.0D).endVertex(); // y
-                        worldRenderer.pos((double) px, (double) py, 0.0D).endVertex(); // z
-
+                        worldRenderer.pos(px, py, 0.0D).endVertex(); // z
                     } else if (side == 3) {
-                        worldRenderer.pos((double) trueX, (double) trueY, 0.0D).endVertex();
-                        worldRenderer.pos((double) px, (double) py, 0.0D).endVertex();
-                        worldRenderer.pos((double) px + 10, (double) py, 0.0D).endVertex();
-
+                        worldRenderer.pos(trueX, trueY, 0.0D).endVertex();
+                        worldRenderer.pos(px, py, 0.0D).endVertex();
+                        worldRenderer.pos((double) px + 10, py, 0.0D).endVertex();
                     } else if (side == 2) {
-                        worldRenderer.pos((double) trueX, (double) trueY, 0.0D).endVertex();
-                        worldRenderer.pos((double) px, (double) py, 0.0D).endVertex();
-                        worldRenderer.pos((double) px + 10, (double) py, 0.0D).endVertex();
-
+                        worldRenderer.pos(trueX, trueY, 0.0D).endVertex();
+                        worldRenderer.pos(px, py, 0.0D).endVertex();
+                        worldRenderer.pos((double) px + 10, py, 0.0D).endVertex();
                     } else {
-                        worldRenderer.pos((double) trueX, (double) trueY, 0.0D).endVertex();
-                        worldRenderer.pos((double) px + 10, (double) py, 0.0D).endVertex();
-                        worldRenderer.pos((double) px, (double) py, 0.0D).endVertex();
+                        worldRenderer.pos(trueX, trueY, 0.0D).endVertex();
+                        worldRenderer.pos((double) px + 10, py, 0.0D).endVertex();
+                        worldRenderer.pos(px, py, 0.0D).endVertex();
                     }
                     renderer.draw();
                     GlStateManager.enableTexture2D();
